wip: transform AABB of objects
Bounding Boxes of objects are calculated once on the BuildTriangles and
never updated so the logic for the collisions don't work as the bbox are
out of place. I think the way to go is to acctually store the
transformation for the model in the scene object then apply it when draw
it and when test the collisions.